requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
numpy<2.0
sentence-transformers
faiss-cpu
transformers
diffusers
ollama
PyPDF2
pandas
openpyxl
langchain>=0.0.300
langchain-community
streamlit
streamlit-authenticator
requests